requirements.txt
robotframework==5.0.1
robotframework-requests==0.9.3
